requirements.txt
falcon==3.1.1
gunicorn==20.1.0
python-dotenv==0.20.0
requests==2.27.1